server/requirements.txt
sanic
mysql-connector-python
requests
simple-crypt